Fix memory protection error I was having when running examples on CentOS.
When I run the `hello-world` application, for example, by running the
binary `woofc-namespace-platform`, the following error message occurs:

"/cspot/woofc-container: error while loading shared libraries: cannot
apply additional memory protection after relocation: Permission denied"

After some Googling, I found various sources[1][2][3] that say this can be
alleviated by adding the ":z" flag to the volume option in `docker run`,
which is used in `woofc-host.c`. After adding the flag and recomipiling, I
can run the examples like normal. When I then remove the flag,
recompile, and re-run the example, using the same, i.e. undeleted, namespace
directory "/cspot/", the examples also run find. So it seems that docker
only needs that flag to be present when docker run is first used to
create the volume /cspot/.

I tried this is two CentOS installations, and the ':z' option was
needed.
It wasn't needed in Ubuntu, but still runs fine with it (in my limited
testing of it in the hello-world and senspot applications).

I'm not an expert on memory protection issues that arise in Linux, so
there may be a better/more targeted fix for this than changing the
`docker run` flag (since it works on Ubuntu, and I'm assuming Rich's
verions of CentOS, without it).

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=472529
[2] https://stackoverflow.com/questions/35218194/what-is-z-flag-in-docker-containers-volumes-from-option
[3] https://github.com/rhatdan/moby1/blob/e6473011583967df4aa5a62f173fb421cae2bb1e/docs/sources/reference/commandline/cli.md